react-router-dom version changed-*read description
Changed the version of react-router-dom in package.json from 6.1.4 to 5.3.0.
This was why our routing wasn't working.
There are a few changes in the latest version and I think we can update them later on.
Then I ran npm-install again to update the necessary changes as part of the version change. This fixed the Issue.

Note: Run npm-install after you pull down the code